requirements.txt
Flask
Flask-SQLAlchemy
azure-storage-blob
Werkzeug
pyodbc